organized scss files
