+++ src/AboutMe.jsx
@@ -1,4 +1,4 @@
-import './About.scss';
+import './scss/About.scss';
 import Desktop from './imgs/desktop3.png'
 
 const About = () => {
